Added get_day_of_week() and get_month_name(), increased FORECAST_DAYS to 8
get_day_of_week() receives a date in the form of a string and returns the name of the day of the week it represents. i.e. "2020-07-14" returns "Tuesday"

get_month_name() receives a dat in the form a string and returns the name of the month, i.e. "January"

Increased FORECAST_DAYS to 8 , as forecast includes current day(today) and 7 days after it, meaning we obtain a total of 8 days worth of forecast data from the API.